LPS-194841 Merger class PathInterpreter with class AMImageRequestHandler, and apply to usages
